Explicitly cast, quiets warnings.

+++ Ball.cpp
@@ -20,9 +20,9 @@ void Ball::render(sf::RenderWindow& window) {
 void Ball::reset(sf::RenderWindow& window) {
 	//Want to have the ball only have specific speeds (Not zero)
 	static int table[] = {-3, -2, -1, 1, 2, 3};
-	position.x = (window.getSize().x)/2;
-	position.y = (window.getSize().y)/2;
-	velocity.x = table[rand() %(sizeof table/ sizeof *table)];
-	velocity.y = table[rand() %(sizeof table/ sizeof *table)];
+	position.x = float((window.getSize().x)/2);
+	position.y = float((window.getSize().y)/2);
+	velocity.x = float(table[rand() %(sizeof table/ sizeof *table)]);
+	velocity.y = float(table[rand() %(sizeof table/ sizeof *table)]);
 }
